Drop the duplicated action modules left behind by the file move
i4 moved nine modules from datasets/action/ into datasets/action/utils/, so the
release now writes them to cosmos_framework/data/generator/action/utils/. The
release only writes destinations, it never removes a path that dropped out of
the mapping, so the pre-move copies survived at the old location and this PR
ships both.

They are not harmless leftovers. The 30 importers still name the old paths, so
the package runs on the stale copies while the mapped ones sit unused, and the
two have already diverged: transforms_test gained a whole test case, and
transforms, json_formatter, action_spec and action_processing all differ. Only
viewpoint_utils and domain_utils are still byte-identical. Any future i4 fix
lands solely in the mapped copy, invisible to the code that actually imports.

The drift gate cannot see this: every import resolves, because both copies
exist. It checks resolvability, not identity.

Repoint the importers at action.utils.* and delete the nine orphans. The six
flat modules that are still mapped there (agibot_fk, libero_pose_utils, the
datasets/ tree, ...) are untouched.

Verified: every action/utils/ file matches forward(i4) exactly while none of the
orphans did, so the mapped copies are authoritative; no old-path imports remain;
all files parse; and tests/nano_inference_smoke_test.py::test_nano_inference_omni
passes end to end, exercising the action policy and forward_dynamics paths that
these imports feed.
