+++ cosmos_framework/data/generator/action/datasets/agibotworld_beta_lerobot_dataset.py
@@ -19,9 +19,9 @@ from cosmos_framework.data.generator.action.agibot_fk import (
     compute_fk_transforms_batch,
     convert_gripper_state_to_open_fraction,
 )
-from cosmos_framework.data.generator.action.action_spec import ActionSpec, Gripper, Pos, Rot, build_action_spec
+from cosmos_framework.data.generator.action.utils.action_spec import ActionSpec, Gripper, Pos, Rot, build_action_spec
 from cosmos_framework.data.generator.action.datasets.base_dataset import ActionBaseDataset
-from cosmos_framework.data.generator.action.pose_utils import pose_abs_to_rel
+from cosmos_framework.data.generator.action.utils.pose_utils import pose_abs_to_rel
 
 PoseConvention = Literal["backward_framewise"]
 Viewpoint = Literal["concat_view", "ego_view"]
