+++ cosmos_framework/data/generator/action/datasets/base_dataset.py
@@ -18,9 +18,9 @@ import torch
 from torch.utils.data import Dataset
 
 from cosmos_framework.data.generator.action.action_normalization import load_action_stats, normalize_action
-from cosmos_framework.data.generator.action.action_spec import ActionSpec
-from cosmos_framework.data.generator.action.domain_utils import get_domain_id
-from cosmos_framework.data.generator.action.pose_utils import compute_idle_frames
+from cosmos_framework.data.generator.action.utils.action_spec import ActionSpec
+from cosmos_framework.data.generator.action.utils.domain_utils import get_domain_id
+from cosmos_framework.data.generator.action.utils.pose_utils import compute_idle_frames
 
 _MODE_CHOICES = ("forward_dynamics", "inverse_dynamics", "wam")
