+++ cosmos_framework/inference/action.py
@@ -9,18 +9,18 @@ from typing import Any
 
 import torch
 
-from cosmos_framework.data.generator.action.action_processing import (
+from cosmos_framework.data.generator.action.utils.action_processing import (
     ActionProcessingRecord,
     make_batched_action_processing_fields,
     pad_action_to_max_dim,
 )
-from cosmos_framework.data.generator.action.domain_utils import (
+from cosmos_framework.data.generator.action.utils.domain_utils import (
     EMBODIMENT_TO_DOMAIN_ID,
     EMBODIMENT_TO_RAW_ACTION_DIM,
     get_domain_id,
 )
-from cosmos_framework.data.generator.action.json_formatter import ActionPromptJsonFormatter
-from cosmos_framework.data.generator.action.transforms import (
+from cosmos_framework.data.generator.action.utils.json_formatter import ActionPromptJsonFormatter
+from cosmos_framework.data.generator.action.utils.transforms import (
     build_sequence_plan_from_mode,
     find_closest_target_size,
     reflection_pad_to_target,
