+++ cosmos_framework/scripts/action_policy_server_robolab.py
@@ -40,14 +40,14 @@ import torch
 import torch.nn.functional as F
 import tyro
 
-from cosmos_framework.data.generator.action.domain_utils import get_domain_id
-from cosmos_framework.data.generator.action.pose_utils import (
+from cosmos_framework.data.generator.action.utils.domain_utils import get_domain_id
+from cosmos_framework.data.generator.action.utils.pose_utils import (
     build_abs_pose_from_components,
     convert_rotation,
     pose_abs_to_rel,
     pose_rel_to_abs,
 )
-from cosmos_framework.data.generator.action.transforms import ActionTransformPipeline
+from cosmos_framework.data.generator.action.utils.transforms import ActionTransformPipeline
 from cosmos_framework.data.generator.joint_dataloader import IterativeJointDataLoader
 from cosmos_framework.inference.args import OmniSetupArgs, OmniSetupOverrides
 from cosmos_framework.inference.common.args import ConfigFileType, ConfigOverrides, tyro_cli
